+++ src/test/java/com/jimmy/diningreviewapi/TestConfig.java
@@ -1,7 +1,7 @@
 package com.jimmy.diningreviewapi;
 
 import com.jimmy.diningreviewapi.repository.querydsl.MemberRepositoryCustom;
-import com.jimmy.diningreviewapi.repository.querydsl.MemberRepositoryCustomImpl;
+import com.jimmy.diningreviewapi.repository.querydsl.impl.MemberRepositoryCustomImpl;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
